Moved recombination map tests to own file.
